make connection modes public
Refactor connection modes to become public for reuse with other modes
