+++ outputs/lumberjack/lumberjack.go
@@ -11,6 +11,7 @@ import (
 	"github.com/elastic/libbeat/common"
 	"github.com/elastic/libbeat/logp"
 	"github.com/elastic/libbeat/outputs"
+	"github.com/elastic/libbeat/outputs/mode"
 )
 
 var debug = logp.MakeDebug("lumberjack")
@@ -35,7 +36,7 @@ func (p lumberjackOutputPlugin) NewOutput(
 }
 
 type lumberjack struct {
-	mode ConnectionMode
+	mode mode.ConnectionMode
 }
 
 const (
@@ -63,7 +64,7 @@ func (lj *lumberjack) init(
 		timeout = time.Duration(config.Timeout) * time.Second
 	}
 
-	var clients []ProtocolClient
+	var clients []mode.ProtocolClient
 	var err error
 	if useTLS {
 		var tlsConfig *tls.Config
@@ -91,22 +92,23 @@ func (lj *lumberjack) init(
 		sendRetries = *config.Max_retries
 	}
 
-	var mode ConnectionMode
+	var m mode.ConnectionMode
 	if len(clients) == 1 {
-		mode, err = newSingleConnectionMode(clients[0], sendRetries, waitRetry, timeout)
+		m, err = mode.NewSingleConnectionMode(clients[0],
+			sendRetries, waitRetry, timeout)
 	} else {
 		loadBalance := config.LoadBalance == nil || *config.LoadBalance
 		if loadBalance {
-			mode, err = newLoadBalancerMode(clients, sendRetries, waitRetry, timeout)
+			m, err = mode.NewLoadBalancerMode(clients, sendRetries, waitRetry, timeout)
 		} else {
-			mode, err = newFailOverConnectionMode(clients, sendRetries, waitRetry, timeout)
+			m, err = mode.NewFailOverConnectionMode(clients, sendRetries, waitRetry, timeout)
 		}
 	}
 	if err != nil {
 		return err
 	}
 
-	lj.mode = mode
+	lj.mode = m
 	return nil
 }
 
@@ -114,10 +116,10 @@ func makeClients(
 	config outputs.MothershipConfig,
 	timeout time.Duration,
 	newTransp func(string) (TransportClient, error),
-) ([]ProtocolClient, error) {
+) ([]mode.ProtocolClient, error) {
 	switch {
 	case len(config.Hosts) > 0:
-		var clients []ProtocolClient
+		var clients []mode.ProtocolClient
 		for _, host := range config.Hosts {
 			transp, err := newTransp(host)
 			if err != nil {
@@ -135,7 +137,7 @@ func makeClients(
 		if err != nil {
 			return nil, err
 		}
-		return []ProtocolClient{newLumberjackClient(transp, timeout)}, nil
+		return []mode.ProtocolClient{newLumberjackClient(transp, timeout)}, nil
 	default:
 		return nil, ErrNoHostsConfigured
 	}
